# docs/swagger/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI QA Payments Lab",
    "version": "1.0.0"
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Healthcheck",
        "operationId": "healthcheck_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Healthcheck Health Get"
                }
              }
            }
          }
        }
      }
    },
    "/transfers": {
      "post": {
        "summary": "Create Transfer Endpoint",
        "operationId": "create_transfer_endpoint_transfers_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTransferRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransferResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/transfers/{transfer_id}": {
      "get": {
        "summary": "Get Transfer Endpoint",
        "operationId": "get_transfer_endpoint_transfers__transfer_id__get",
        "parameters": [
          {
            "name": "transfer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Transfer Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransferResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CreateTransferRequest": {
        "properties": {
          "from_account_id": {
            "type": "string",
            "minLength": 1,
            "title": "From Account Id"
          },
          "to_alias": {
            "type": "string",
            "minLength": 3,
            "title": "To Alias"
          },
          "amount": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Amount"
          },
          "currency": {
            "type": "string",
            "maxLength": 3,
            "minLength": 3,
            "title": "Currency"
          },
          "idempotency_key": {
            "type": "string",
            "minLength": 3,
            "title": "Idempotency Key"
          }
        },
        "type": "object",
        "required": [
          "from_account_id",
          "to_alias",
          "amount",
          "currency",
          "idempotency_key"
        ],
        "title": "CreateTransferRequest"
      },
      "ErrorResponse": {
        "properties": {
          "error_code": {
            "type": "string",
            "title": "Error Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "error_code",
          "message"
        ],
        "title": "ErrorResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "TransferResponse": {
        "properties": {
          "transfer_id": {
            "type": "string",
            "title": "Transfer Id"
          },
          "state": {
            "$ref": "#/components/schemas/TransferState"
          },
          "amount": {
            "type": "number",
            "title": "Amount"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "from_account_id": {
            "type": "string",
            "title": "From Account Id"
          },
          "to_alias": {
            "type": "string",
            "title": "To Alias"
          },
          "idempotency_key": {
            "type": "string",
            "title": "Idempotency Key"
          }
        },
        "type": "object",
        "required": [
          "transfer_id",
          "state",
          "amount",
          "currency",
          "from_account_id",
          "to_alias",
          "idempotency_key"
        ],
        "title": "TransferResponse"
      },
      "TransferState": {
        "type": "string",
        "enum": [
          "PENDING",
          "COMPLETED",
          "CANCELLED"
        ],
        "title": "TransferState"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}